Index.js added to the proyect
